chore: refactor the code
