+++ spa/src/utils/dynamicTableHelper.tsx
@@ -5,98 +5,7 @@ import { token } from "@atlaskit/tokens";
 import Lozenge from "@atlaskit/lozenge";
 import EditIcon from "@atlaskit/icon/glyph/edit";
 import MoreIcon from "@atlaskit/icon/glyph/more";
-
-export type Row = {
-	key: string;
-	isHighlighted: boolean;
-	cells: { key: string | number; content: JSX.Element | string | number }[];
-};
-
-export type Account = {
-	login: string;
-	id: number;
-	avatar_url: string;
-	type?: string;
-	site_admin?: boolean;
-};
-
-export type SuccessfulConnection = {
-	app_slug: string;
-	syncWarning: string;
-	id: number;
-	account: Account;
-	repository_selection: string;
-	app_id: number;
-	target_id: number;
-	target_type: string;
-	created_at: string;
-	updated_at: string;
-	syncStatus: string;
-	totalNumberOfRepos: number;
-	numberOfSyncedRepos: number;
-	jiraHost: string;
-	isGlobalInstall: boolean;
-	backfillSince: string | null;
-};
-
-export type FailedCloudConnection = {
-	id: number;
-	deleted: boolean;
-	orgName?: string;
-};
-
-export type GhCloudSubscriptions = {
-	successfulCloudConnections: SuccessfulConnection[];
-	failedCloudConnections: FailedCloudConnection[];
-};
-
-export type FailedConnection = {
-	id: number;
-	deleted: boolean;
-	orgName?: string;
-};
-
-export type Installation = {
-	id: number;
-	account: Account;
-	target_type: string;
-	created_at: string;
-	updated_at: string;
-	syncStatus: string;
-	totalNumberOfRepos: number;
-	numberOfSyncedRepos: number;
-	backfillSince: null | string;
-	jiraHost: string;
-};
-
-export type GitHubEnterpriseApplication = {
-	id: number;
-	uuid: string;
-	appId: number;
-	gitHubBaseUrl: string;
-	gitHubClientId: string;
-	gitHubAppName: string;
-	installationId: number;
-	createdAt: string;
-	updatedAt: string;
-	successfulConnections: SuccessfulConnection[];
-	failedConnections: FailedConnection[];
-	installations: {
-		fulfilled: Installation[];
-		rejected: any[];
-		total: number;
-	};
-};
-
-export type GhEnterpriseServer = {
-	gitHubBaseUrl: string;
-	applications: GitHubEnterpriseApplication[];
-};
-
-export type GHSUbscriptions = {
-	ghCloudSubscriptions: GhCloudSubscriptions;
-	ghEnterpriseServers: GhEnterpriseServer[];
-};
+import { SuccessfulConnection, Row } from "../rest-interfaces";
 
 const RowWrapper = styled.div`
 	display: flex;
