+++ src/rest-interfaces/index.ts
@@ -1,4 +1,3 @@
-
 export type GetRedirectUrlResponse = {
 	redirectUrl: string;
 	state: string;
@@ -16,11 +15,11 @@ export type UsersGetAuthenticatedResponse = {
 
 export type GetGitHubAppsUrlResponse = {
 	appInstallationUrl: string;
-}
+};
 
 export type OrganizationsResponse = {
 	orgs: Array<GitHubInstallationType>;
-}
+};
 
 export type GitHubInstallationType = {
 	account: GitHubInstallationAccountType;
@@ -49,7 +48,7 @@ export type ErrorType = "warning" | "error";
 export type ApiError = {
 	message: string;
 	errorCode: ErrorCode;
-}
+};
 
 //DO NOT USE ENUM as webpack can't handler anything none "type"
 //ts-load is not a real typescript compile, it only strips off the types, hence cannot process Enum/Class/etc
@@ -63,3 +62,95 @@ export type ErrorCode =
 	| "SSO_LOGIN"
 	| "RESOURCE_NOT_FOUND"
 	| "UNKNOWN";
+
+export type Row = {
+	key: string;
+	isHighlighted: boolean;
+	cells: { key: string | number; content: JSX.Element | string | number }[];
+};
+
+export type Account = {
+	login: string;
+	id: number;
+	avatar_url: string;
+	type?: string;
+	site_admin?: boolean;
+};
+
+export type SuccessfulConnection = {
+	app_slug: string;
+	syncWarning: string;
+	id: number;
+	account: Account;
+	repository_selection: string;
+	app_id: number;
+	target_id: number;
+	target_type: string;
+	created_at: string;
+	updated_at: string;
+	syncStatus: string;
+	totalNumberOfRepos: number;
+	numberOfSyncedRepos: number;
+	jiraHost: string;
+	isGlobalInstall: boolean;
+	backfillSince: string | null;
+};
+
+export type FailedCloudConnection = {
+	id: number;
+	deleted: boolean;
+	orgName?: string;
+};
+
+export type GhCloudSubscriptions = {
+	successfulCloudConnections: SuccessfulConnection[];
+	failedCloudConnections: FailedCloudConnection[];
+};
+
+export type FailedConnection = {
+	id: number;
+	deleted: boolean;
+	orgName?: string;
+};
+
+export type Installation = {
+	id: number;
+	account: Account;
+	target_type: string;
+	created_at: string;
+	updated_at: string;
+	syncStatus: string;
+	totalNumberOfRepos: number;
+	numberOfSyncedRepos: number;
+	backfillSince: null | string;
+	jiraHost: string;
+};
+
+export type GitHubEnterpriseApplication = {
+	id: number;
+	uuid: string;
+	appId: number;
+	gitHubBaseUrl: string;
+	gitHubClientId: string;
+	gitHubAppName: string;
+	installationId: number;
+	createdAt: string;
+	updatedAt: string;
+	successfulConnections: SuccessfulConnection[];
+	failedConnections: FailedConnection[];
+	installations: {
+		fulfilled: Installation[];
+		rejected: any[];
+		total: number;
+	};
+};
+
+export type GhEnterpriseServer = {
+	gitHubBaseUrl: string;
+	applications: GitHubEnterpriseApplication[];
+};
+
+export type GHSUbscriptions = {
+	ghCloudSubscriptions: GhCloudSubscriptions;
+	ghEnterpriseServers: GhEnterpriseServer[];
+};
